修复金额转Long类型后的错误

+++ pay-service/pay-api/pay-charge-api/src/main/java/com/gloryjie/pay/trade/dto/ChargeDto.java
@@ -48,9 +48,9 @@ public class ChargeDto {
     /**
      * 支付单金额
      */
-    private BigDecimal amount;
+    private Long amount;
 
-    private BigDecimal actualAmount;
+    private Long actualAmount;
 
     private ChargeStatus status;
